kafka/tests: Moved functions to common fixture
Will be used in future testing of other kafka messages
