+++ src/v/kafka/server/tests/metadata_test.cc
@@ -7,30 +7,22 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "cluster/config_frontend.h"
 #include "cluster/security_frontend.h"
 #include "kafka/client/transport.h"
 #include "kafka/protocol/create_topics.h"
 #include "kafka/protocol/metadata.h"
-#include "kafka/protocol/sasl_authenticate.h"
-#include "kafka/protocol/sasl_handshake.h"
 #include "kafka/protocol/types.h"
 #include "kafka/server/handlers/details/security.h"
 #include "kafka/server/handlers/metadata.h"
-#include "model/ktp.h"
 #include "model/timeout_clock.h"
-#include "random/generators.h"
 #include "redpanda/tests/fixture.h"
 #include "security/acl.h"
 #include "security/scram_algorithm.h"
-#include "security/scram_authenticator.h"
 #include "security/types.h"
 #include "test_utils/random_bytes.h"
 
 #include <absl/algorithm/container.h>
 #include <boost/test/tools/old/interface.hpp>
-
-#include <chrono>
 
 static const int32_t not_provided_authz_return = -2147483648;
 static const std::vector<security::acl_operation> default_cluster_auths = {
@@ -86,95 +78,6 @@ protected:
                        user, credential, model::timeout_clock::now() + 5s)
                      .get();
         BOOST_REQUIRE_EQUAL(err, cluster::errc::success);
-    }
-
-    void enable_sasl() {
-        cluster::config_update_request r{.upsert = {{"enable_sasl", "true"}}};
-        auto res = app.controller->get_config_frontend()
-                     .local()
-                     .patch(r, model::timeout_clock::now() + 1s)
-                     .get();
-        BOOST_REQUIRE(!res.errc);
-    }
-
-    void disable_sasl() {
-        cluster::config_update_request r{.upsert = {{"enable_sasl", "false"}}};
-        auto res = app.controller->get_config_frontend()
-                     .local()
-                     .patch(r, model::timeout_clock::now() + 1s)
-                     .get();
-        BOOST_REQUIRE(!res.errc);
-    }
-
-    security::server_first_message send_scram_client_first(
-      kafka::client::transport& client,
-      const security::client_first_message& client_first) {
-        kafka::sasl_authenticate_request client_first_req;
-        {
-            auto msg = client_first.message();
-            client_first_req.data.auth_bytes = bytes(msg.cbegin(), msg.cend());
-        }
-        auto client_first_resp = client.dispatch(client_first_req).get();
-        BOOST_REQUIRE_EQUAL(
-          client_first_resp.data.error_code, kafka::error_code::none);
-        return security::server_first_message(
-          client_first_resp.data.auth_bytes);
-    }
-
-    security::server_final_message send_scram_client_final(
-      kafka::client::transport& client,
-      const security::client_final_message& client_final) {
-        kafka::sasl_authenticate_request client_last_req;
-        {
-            auto msg = client_final.message();
-            client_last_req.data.auth_bytes = bytes(msg.cbegin(), msg.cend());
-        }
-        auto client_last_resp = client.dispatch(client_last_req).get();
-
-        BOOST_REQUIRE_EQUAL(
-          client_last_resp.data.error_code, kafka::error_code::none);
-        return security::server_final_message(
-          std::move(client_last_resp.data.auth_bytes));
-    }
-
-    void do_sasl_handshake(kafka::client::transport& client) {
-        kafka::sasl_handshake_request req;
-        req.data.mechanism = security::scram_sha256_authenticator::name;
-
-        auto resp = client.dispatch(req).get();
-        BOOST_REQUIRE_EQUAL(resp.data.error_code, kafka::error_code::none);
-    }
-
-    void authn_kafka_client(
-      kafka::client::transport& client,
-      const ss::sstring& username,
-      const ss::sstring& password) {
-        do_sasl_handshake(client);
-        const auto nonce = random_generators::gen_alphanum_string(130);
-        const security::client_first_message client_first(username, nonce);
-        const auto server_first = send_scram_client_first(client, client_first);
-
-        BOOST_REQUIRE(
-          std::string_view(server_first.nonce()).starts_with(nonce));
-        BOOST_REQUIRE_GE(
-          server_first.iterations(), security::scram_sha256::min_iterations);
-        security::client_final_message client_final(
-          bytes::from_string("n,,"), server_first.nonce());
-        auto salted_password = security::scram_sha256::hi(
-          bytes(password.cbegin(), password.cend()),
-          server_first.salt(),
-          server_first.iterations());
-        client_final.set_proof(security::scram_sha256::client_proof(
-          salted_password, client_first, server_first, client_final));
-
-        auto server_final = send_scram_client_final(client, client_final);
-        BOOST_REQUIRE(!server_final.error());
-
-        auto server_key = security::scram_sha256::server_key(salted_password);
-        auto server_sig = security::scram_sha256::server_signature(
-          server_key, client_first, server_first, client_final);
-
-        BOOST_REQUIRE_EQUAL(server_final.signature(), server_sig);
     }
 };
